feature: create nativeCli source set for test
